chore: rename authConfig.js to auth.js
