+++ code/workspaces/web-app/src/app.js
@@ -5,7 +5,7 @@ import { ThemeProvider as MuiThemeProvider } from '@material-ui/core/styles';
 import configureStore from './store/configureStore';
 import theme from './theme';
 import Router from './router';
-import { initialiseAuth } from './config/authConfig';
+import { initialiseAuth } from './config/auth';
 import { initialiseVersion } from './config/version';
 
 async function createApplication() {
