+++ code/workspaces/web-app/src/components/app/RequireAuth.spec.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { createShallow } from '@material-ui/core/test-utils';
-import { getAuth } from '../../config/authConfig';
+import { getAuth } from '../../config/auth';
 import { useCurrentUserPermissions, useCurrentUserTokens } from '../../hooks/authHooks';
 import RequireAuth, { effectFn } from './RequireAuth';
 import authActions from '../../actions/authActions';
 
 jest.mock('react-redux');
-jest.mock('../../config/authConfig');
+jest.mock('../../config/auth');
 jest.mock('../../hooks/authHooks');
 jest.mock('../../actions/authActions');
