+++ code/workspaces/web-app/src/components/app/UserMenu.js
@@ -4,7 +4,7 @@ import Avatar from '@material-ui/core/Avatar';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import { withStyles } from '@material-ui/core/styles';
-import { getAuth } from '../../config/authConfig';
+import { getAuth } from '../../config/auth';
 
 const styles = theme => ({
   container: {
