+++ code/workspaces/web-app/src/components/app/UserMenu.spec.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { createShallow } from '@material-ui/core/test-utils';
 import UserMenu from './UserMenu';
-import { getAuth } from '../../config/authConfig';
+import { getAuth } from '../../config/auth';
 
-jest.mock('../../config/authConfig');
+jest.mock('../../config/auth');
 const logout = jest.fn();
 getAuth.mockImplementation(() => ({
   logout,
