+++ code/workspaces/web-app/src/components/welcome/VerifyEmail.spec.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { createShallow } from '@material-ui/core/test-utils';
 import VerifyEmail from './VerifyEmail';
-import { getAuth } from '../../config/authConfig';
+import { getAuth } from '../../config/auth';
 
-jest.mock('../../config/authConfig');
+jest.mock('../../config/auth');
 getAuth.mockImplementation(() => ({
   login: jest.fn().mockName('login'),
 }));
