+++ code/workspaces/web-app/src/config/auth.js
