+++ code/workspaces/web-app/src/config/auth.spec.js
@@ -1,7 +1,7 @@
 import MockAdapter from 'axios-mock-adapter';
 import axios from 'axios';
 import data from 'common/src/config/web_auth_config.json';
-import { initialiseAuth, getAuth } from './authConfig';
+import { initialiseAuth, getAuth } from './auth';
 
 const httpMock = new MockAdapter(axios);
 httpMock.onGet('/web_auth_config.json').reply(() => [200, data]);
